refactor: Refactor mock files and add data folder
